+++ .streamlit/config.toml
@@ -1,0 +1,6 @@
+[theme]
+primaryColor = "#5ebdfa"
+backgroundColor = "#002e4d"
+secondaryBackgroundColor = "#004978"
+textColor = "#ffffff"
+font = "sans serif"
