+++ .streamlit/secrets.toml
@@ -1,0 +1,2 @@
+OPENAI_API_KEY = "dsafsadsad API" 
+#örnektir
