+++ streamlit_app.py
@@ -7,21 +7,21 @@ import webbrowser
 import requests
 
 giris_page = st.Page(
-    page = "app.py",
+    page = "✋_app.py",
     title = "Teknik Mülakata Giriş",
     icon = ":material/account_circle:",
     default = True,    
 )
 
 llama_page = st.Page(
-    page = "llama.py",
+    page = "🦙_llama.py",
     title = "llama",
     icon = ":material/account_circle:",
     default = True,    
 )
 
 mistral_page = st.Page(
-    page = "mistral.py",
+    page = "🤖_mistral.py",
     title = "mistral",
     icon = ":material/account_circle:",
     default = True,    
@@ -29,14 +29,14 @@ mistral_page = st.Page(
 
 
 gpt_page = st.Page(
-    page = "chatgpt.py",
+    page = "⚛️_chatgpt.py",
     title = "ChatGPT",
     icon = ":material/account_circle:",
     default = True,    
 )
 
 mulakat_page = st.Page(
-    page = "mulakat.py",
+    page = "🗣️_mulakat.py",
     title = "Karşılıklı Mülakat",
     icon = ":material/account_circle:",
     default = True,    
@@ -44,7 +44,7 @@ mulakat_page = st.Page(
 
 
 
-pg = st.navigation(pages=[st.Page("app.py"), st.Page("llama.py"), st.Page("mistral.py"), 
-                          st.Page("chatgpt.py"), st.Page("mulakat.py")])
+pg = st.navigation(pages=[st.Page("✋_app.py"), st.Page("🦙_llama.py"), st.Page("🤖_mistral.py"), 
+                          st.Page("⚛️_chatgpt.py"), st.Page("🗣️_mulakat.py")])
 
 pg.run()
