+++ 🗣️_mulakat.py
@@ -1,0 +1,52 @@
+import streamlit as st
+import pandas as pd
+import numpy as np
+from transformers import pipeline
+import time
+import webbrowser
+import requests
+
+st.title("Karşılıklı Mülakat")
+
+
+def open_main():
+    url = 'http://localhost:8502'  
+    webbrowser.open_new_tab(url)
+
+
+from openai import OpenAI
+import streamlit as st
+
+
+client = OpenAI(api_key=st.secrets["OPENAI_API_KEY"])  #API tanımlı değil. Örnek olsun diye konulmustur
+
+if "openai_model" not in st.session_state:
+    st.session_state["openai_model"] = "gpt-3.5-turbo"
+
+if "messages" not in st.session_state:
+    st.session_state.messages = []
+
+for message in st.session_state.messages:
+    with st.chat_message(message["role"]):
+        st.markdown(message["content"])
+
+if prompt := st.chat_input("What is up?"):
+    st.session_state.messages.append({"role": "user", "content": prompt})
+    with st.chat_message("user"):
+        st.markdown(prompt)
+
+    with st.chat_message("assistant"):
+        stream = client.chat.completions.create(
+            model=st.session_state["openai_model"],
+            messages=[
+                {"role": m["role"], "content": m["content"]}
+                for m in st.session_state.messages
+            ],
+            stream=True,
+        )
+        response = st.write_stream(stream)
+    st.session_state.messages.append({"role": "assistant", "content": response})
+
+
+
+st.button("Ana Sayfaya Dön", on_click=open_main)
